Changed dataclass state to state_dict for better extensibility and added SettingPosition
