+++ deep_conmech/simulator/setting/setting_forces.py
@@ -1,6 +1,7 @@
 from conmech.helpers import nph
 from conmech.solvers.optimization.schur_complement import SchurComplement
 from deep_conmech.simulator.dynamics.dynamics import Dynamics
+from deep_conmech.simulator.setting.setting_position import SettingPosition
 
 
 def L2_new(a, C, E):
@@ -10,7 +11,7 @@ def L2_new(a, C, E):
     return value
 
 
-class SettingForces(Dynamics):
+class SettingForces(SettingPosition):
     def __init__(
             self, mesh_data, body_prop, schedule, normalize_by_rotation: bool, create_in_subprocess,
     ):
